Rename TagsCloudDrawerModuleSettings to TagsCloudVisualisationSettings and bind it to project, but not module of di-container
